DPL Analysis: Rearrange code in AnalysisTask.h
