+++ Framework/Core/src/AnalysisManagers.h
@@ -1,0 +1,399 @@
+// Copyright CERN and copyright holders of ALICE O2. This software is
+// distributed under the terms of the GNU General Public License v3 (GPL
+// Version 3), copied verbatim in the file "COPYING".
+//
+// See http://alice-o2.web.cern.ch/license for full licensing information.
+//
+// In applying this license CERN does not waive the privileges and immunities
+// granted to it by virtue of its status as an Intergovernmental Organization
+// or submit itself to any jurisdiction.
+
+#ifndef FRAMEWORK_ANALYSISMANAGERS_H
+#define FRAMEWORK_ANALYSISMANAGERS_H
+#include "Framework/AnalysisHelpers.h"
+#include "Framework/Kernels.h"
+#include "Framework/ASoA.h"
+#include "Framework/ProcessingContext.h"
+#include "Framework/EndOfStreamContext.h"
+#include "Framework/HistogramRegistry.h"
+#include "Framework/ConfigParamSpec.h"
+#include "Framework/ConfigParamRegistry.h"
+#include "Framework/InitContext.h"
+#include "Framework/RootConfigParamHelpers.h"
+
+namespace o2::framework
+{
+
+template <typename ANY>
+struct PartitionManager {
+  template <typename... T2s>
+  static void setPartition(ANY&, T2s&...)
+  {
+  }
+
+  template <typename... Ts>
+  static void bindExternalIndices(ANY&, Ts*...)
+  {
+  }
+
+  template <typename... Ts>
+  static void getBoundToExternalIndices(ANY&, Ts&...)
+  {
+  }
+};
+
+template <typename T>
+struct PartitionManager<Partition<T>> {
+  template <typename T2>
+  static void doSetPartition(Partition<T>& partition, T2& table)
+  {
+    if constexpr (std::is_same_v<T, T2>) {
+      partition.setTable(table);
+    }
+  }
+
+  template <typename... T2s>
+  static void setPartition(Partition<T>& partition, T2s&... tables)
+  {
+    (doSetPartition(partition, tables), ...);
+  }
+
+  template <typename... Ts>
+  static void bindExternalIndices(Partition<T>& partition, Ts*... tables)
+  {
+    partition.bindExternalIndices(tables...);
+  }
+
+  template <typename... Ts>
+  static void getBoundToExternalIndices(Partition<T>& partition, Ts&... tables)
+  {
+    partition.getBoundToExternalIndices(tables...);
+  }
+};
+
+template <typename ANY>
+struct FilterManager {
+  static bool createExpressionTrees(ANY&, std::vector<ExpressionInfo>&)
+  {
+    return false;
+  }
+};
+
+template <>
+struct FilterManager<expressions::Filter> {
+  static bool createExpressionTrees(expressions::Filter const& filter, std::vector<ExpressionInfo>& expressionInfos)
+  {
+    updateExpressionInfos(filter, expressionInfos);
+    return true;
+  }
+};
+
+/// SFINAE placeholder
+template <typename T>
+struct OutputManager {
+  template <typename ANY>
+  static bool appendOutput(std::vector<OutputSpec>&, ANY&, uint32_t)
+  {
+    return false;
+  }
+
+  template <typename ANY>
+  static bool prepare(ProcessingContext&, ANY&)
+  {
+    return false;
+  }
+
+  template <typename ANY>
+  static bool postRun(EndOfStreamContext&, ANY&)
+  {
+    return true;
+  }
+
+  template <typename ANY>
+  static bool finalize(ProcessingContext&, ANY&)
+  {
+    return true;
+  }
+};
+
+/// Produces specialization
+template <typename TABLE>
+struct OutputManager<Produces<TABLE>> {
+  static bool appendOutput(std::vector<OutputSpec>& outputs, Produces<TABLE>& what, uint32_t)
+  {
+    outputs.emplace_back(what.spec());
+    return true;
+  }
+  static bool prepare(ProcessingContext& context, Produces<TABLE>& what)
+  {
+    what.resetCursor(context.outputs().make<TableBuilder>(what.ref()));
+    return true;
+  }
+  static bool finalize(ProcessingContext&, Produces<TABLE>&)
+  {
+    return true;
+  }
+  static bool postRun(EndOfStreamContext&, Produces<TABLE>&)
+  {
+    return true;
+  }
+};
+
+/// HistogramRegistry specialization
+template <>
+struct OutputManager<HistogramRegistry> {
+  static bool appendOutput(std::vector<OutputSpec>& outputs, HistogramRegistry& what, uint32_t)
+  {
+    outputs.emplace_back(what.spec());
+    return true;
+  }
+  static bool prepare(ProcessingContext&, HistogramRegistry&)
+  {
+    return true;
+  }
+
+  static bool finalize(ProcessingContext&, HistogramRegistry&)
+  {
+    return true;
+  }
+
+  static bool postRun(EndOfStreamContext&, HistogramRegistry&)
+  {
+    return true;
+  }
+};
+
+/// OutputObj specialization
+template <typename T>
+struct OutputManager<OutputObj<T>> {
+  static bool appendOutput(std::vector<OutputSpec>& outputs, OutputObj<T>& what, uint32_t hash)
+  {
+    what.setHash(hash);
+    outputs.emplace_back(what.spec());
+    return true;
+  }
+  static bool prepare(ProcessingContext&, OutputObj<T>&)
+  {
+    return true;
+  }
+
+  static bool finalize(ProcessingContext&, OutputObj<T>&)
+  {
+    return true;
+  }
+
+  static bool postRun(EndOfStreamContext& context, OutputObj<T>& what)
+  {
+    context.outputs().snapshot(what.ref(), *what);
+    return true;
+  }
+};
+
+/// Spawns specializations
+template <typename O>
+static auto extractOriginal(ProcessingContext& pc)
+{
+  return pc.inputs().get<TableConsumer>(aod::MetadataTrait<O>::metadata::tableLabel())->asArrowTable();
+}
+
+template <typename... Os>
+static std::vector<std::shared_ptr<arrow::Table>> extractOriginals(framework::pack<Os...>, ProcessingContext& pc)
+{
+  return {extractOriginal<Os>(pc)...};
+}
+
+template <typename T>
+struct OutputManager<Spawns<T>> {
+  static bool appendOutput(std::vector<OutputSpec>& outputs, Spawns<T>& what, uint32_t)
+  {
+    outputs.emplace_back(what.spec());
+    return true;
+  }
+
+  static bool prepare(ProcessingContext& pc, Spawns<T>& what)
+  {
+    using metadata = typename std::decay_t<decltype(what)>::metadata;
+    auto original_table = soa::ArrowHelpers::joinTables(extractOriginals(typename metadata::originals{}, pc));
+    what.table = std::make_shared<T>(o2::soa::spawner(what.pack(), original_table.get()));
+    return true;
+  }
+
+  static bool finalize(ProcessingContext&, Spawns<T>&)
+  {
+    return true;
+  }
+
+  static bool postRun(EndOfStreamContext& eosc, Spawns<T>& what)
+  {
+    using metadata = typename std::decay_t<decltype(what)>::metadata;
+    eosc.outputs().adopt(Output{metadata::origin(), metadata::description()}, what.asArrowTable());
+    return true;
+  }
+};
+
+/// Builds specialization
+template <typename O>
+static auto extractTypedOriginal(ProcessingContext& pc)
+{
+  ///FIXME: this should be done in invokeProcess() as some of the originals may be compound tables
+  return O{pc.inputs().get<TableConsumer>(aod::MetadataTrait<O>::metadata::tableLabel())->asArrowTable()};
+}
+
+template <typename... Os>
+static auto extractOriginalsTuple(framework::pack<Os...>, ProcessingContext& pc)
+{
+  return std::make_tuple(extractTypedOriginal<Os>(pc)...);
+}
+
+template <typename T, typename P>
+struct OutputManager<Builds<T, P>> {
+  static bool appendOutput(std::vector<OutputSpec>& outputs, Builds<T, P>& what, uint32_t)
+  {
+    outputs.emplace_back(what.spec());
+    return true;
+  }
+
+  static bool prepare(ProcessingContext& pc, Builds<T, P>& what)
+  {
+    using metadata = typename std::decay_t<decltype(what)>::metadata;
+    return what.build(typename metadata::index_pack_t{},
+                      extractTypedOriginal<typename metadata::Key>(pc),
+                      extractOriginalsTuple(typename metadata::originals{}, pc));
+  }
+
+  static bool finalize(ProcessingContext&, Builds<T, P>&)
+  {
+    return true;
+  }
+
+  static bool postRun(EndOfStreamContext& eosc, Builds<T, P>& what)
+  {
+    using metadata = typename std::decay_t<decltype(what)>::metadata;
+    eosc.outputs().adopt(Output{metadata::origin(), metadata::description()}, what.asArrowTable());
+    return true;
+  }
+};
+
+template <typename T>
+class has_instance
+{
+  typedef char one;
+  struct two {
+    char x[2];
+  };
+
+  template <typename C>
+  static one test(decltype(&C::instance));
+  template <typename C>
+  static two test(...);
+
+ public:
+  enum { value = sizeof(test<T>(nullptr)) == sizeof(char) };
+};
+
+template <typename T>
+struct ServiceManager {
+  template <typename ANY>
+  static bool prepare(InitContext&, ANY&)
+  {
+    return false;
+  }
+};
+
+template <typename T>
+struct ServiceManager<Service<T>> {
+  static bool prepare(InitContext& context, Service<T>& service)
+  {
+    if constexpr (has_instance<T>::value) {
+      service.service = &(T::instance()); // Sigh...
+      return true;
+    } else {
+      service.service = context.services().get<T>();
+      return true;
+    }
+    return false;
+  }
+};
+
+template <typename T>
+struct OptionManager {
+  template <typename ANY>
+  static bool appendOption(std::vector<ConfigParamSpec>&, ANY&)
+  {
+    return false;
+  }
+
+  template <typename ANY>
+  static bool prepare(InitContext&, ANY&)
+  {
+    return false;
+  }
+};
+
+template <typename T>
+struct OptionManager<Configurable<T>> {
+  static bool appendOption(std::vector<ConfigParamSpec>& options, Configurable<T>& what)
+  {
+    if constexpr (variant_trait_v<typename std::decay<T>::type> != VariantType::Unknown) {
+      options.emplace_back(ConfigParamSpec{what.name, variant_trait_v<typename std::decay<T>::type>, what.value, {what.help}});
+    } else {
+      auto specs = RootConfigParamHelpers::asConfigParamSpecs<T>(what.name, what.value);
+      options.insert(options.end(), specs.begin(), specs.end());
+    }
+    return true;
+  }
+
+  static bool prepare(InitContext& context, Configurable<T>& what)
+  {
+    if constexpr (variant_trait_v<typename std::decay<T>::type> != VariantType::Unknown) {
+      what.value = context.options().get<T>(what.name.c_str());
+    } else {
+      auto pt = context.options().get<boost::property_tree::ptree>(what.name.c_str());
+      what.value = RootConfigParamHelpers::as<T>(pt);
+    }
+    return true;
+  }
+};
+
+/// Manager template to facilitate extended tables spawning
+template <typename T>
+struct SpawnManager {
+  static bool requestInputs(std::vector<InputSpec>&, T const&) { return false; }
+};
+
+template <typename TABLE>
+struct SpawnManager<Spawns<TABLE>> {
+  static bool requestInputs(std::vector<InputSpec>& inputs, Spawns<TABLE>& spawns)
+  {
+    auto base_specs = spawns.base_specs();
+    for (auto& base_spec : base_specs) {
+      if (std::find_if(inputs.begin(), inputs.end(), [&](InputSpec const& spec) { return base_spec.binding == spec.binding; }) == inputs.end()) {
+        inputs.emplace_back(base_spec);
+      }
+    }
+    return true;
+  }
+};
+
+/// Manager template for building index tables
+template <typename T>
+struct IndexManager {
+  static bool requestInputs(std::vector<InputSpec>&, T const&) { return false; };
+};
+
+template <typename IDX, typename P>
+struct IndexManager<Builds<IDX, P>> {
+  static bool requestInputs(std::vector<InputSpec>& inputs, Builds<IDX, P>& builds)
+  {
+    auto base_specs = builds.base_specs();
+    for (auto& base_spec : base_specs) {
+      if (std::find_if(inputs.begin(), inputs.end(), [&](InputSpec const& spec) { return base_spec.binding == spec.binding; }) == inputs.end()) {
+        inputs.emplace_back(base_spec);
+      }
+    }
+    return true;
+  }
+};
+} // namespace o2::framework
+
+#endif // ANALYSISMANAGERS_H
